+++ app.py
@@ -1,15 +1,11 @@
 from flask import Flask, render_template, request, redirect, url_for, jsonify, Response
 from datetime import date, timedelta, datetime
-import csv, io, random
+import csv, io
 import config
 from db import close_db, query, execute
-import os, uuid
-from PIL import Image
-import re
+import os
 from difflib import get_close_matches
-import requests
-
-
+from helpers import *
 
 UPLOAD_DIR = os.path.join(os.path.dirname(__file__), 'static', 'uploads')
 os.makedirs(UPLOAD_DIR, exist_ok=True)
@@ -17,147 +13,6 @@ os.makedirs(UPLOAD_DIR, exist_ok=True)
 app = Flask(__name__)
 app.config['SECRET_KEY'] = config.SECRET_KEY
 app.teardown_appcontext(close_db)
-
-
-def _name_key(s):
-    return re.sub(r'\s+', ' ', (s or '').strip().lower())
-
-def combine_override_results(raw, lib, web):
-    toks = set(ingredient_terms_from_text(raw))
-    for r in lib:
-        r['source'] = 'library'
-        r['score'] = r.get('score', 1)
-    for w in web:
-        w['source'] = 'web'
-        name_tokens = set(re.findall(r'[a-z]+', (w.get('name') or '').lower()))
-        overlap = len(toks & name_tokens)
-        area = (w.get('cuisine') or '').lower()
-        bias = 1 if area in ('pakistani','indian','bangladeshi','afghan') else 0
-        w['score'] = w.get('score', 1) + overlap + bias
-    merged = lib + web
-    out, seen = [], set()
-    for m in merged:
-        k = _name_key(m.get('name'))
-        if k in seen:
-            continue
-        seen.add(k)
-        out.append(m)
-    out.sort(key=lambda x: (-x.get('score',0), 0 if x.get('source')=='library' else 1, x.get('time_min') or 999, x.get('name') or ''))
-    return out
-
-
-def web_find_recipes(raw):
-    q = raw.replace('+',' ').replace('|',' ').replace('-',' ').strip()
-    out = []
-    try:
-        r = requests.get('https://www.themealdb.com/api/json/v1/1/search.php', params={'s': q}, timeout=6)
-        j = r.json() if r.status_code==200 else {}
-        meals = j.get('meals') or []
-        for m in meals:
-            area = (m.get('strArea') or '')
-            if area not in ['Pakistani','Indian','Bangladeshi','Afghan','Middle Eastern','Arabic','Unknown']:
-                continue
-            out.append({
-                'id': int(m.get('idMeal')),
-                'name': m.get('strMeal'),
-                'cuisine': area or '—',
-                'time_min': 40,
-                'difficulty': 'Medium',
-                'veg': 0,
-                'spice_level': 'Medium',
-                'image_url': m.get('strMealThumb'),
-                'external': True,
-                'source_url': m.get('strSource') or f"https://www.themealdb.com/meal/{m.get('idMeal')}"
-            })
-        if not out:
-            toks = [t for t in re.split(r'[\s\+\|\-]+', q.lower()) if t]
-            best = toks[0] if toks else 'dal'
-            r = requests.get('https://www.themealdb.com/api/json/v1/1/search.php', params={'s': best}, timeout=6)
-            j = r.json() if r.status_code==200 else {}
-            meals = j.get('meals') or []
-            for m in meals[:8]:
-                area = (m.get('strArea') or '')
-                out.append({
-                    'id': int(m.get('idMeal')),
-                    'name': m.get('strMeal'),
-                    'cuisine': area or '—',
-                    'time_min': 40,
-                    'difficulty': 'Medium',
-                    'veg': 0,
-                    'spice_level': 'Medium',
-                    'image_url': m.get('strMealThumb'),
-                    'external': True,
-                    'source_url': m.get('strSource') or f"https://www.themealdb.com/meal/{m.get('idMeal')}"
-                })
-    except Exception:
-        return []
-    names = set()
-    dedup = []
-    for r in out:
-        if r['name'] in names: continue
-        names.add(r['name']); dedup.append(r)
-    return dedup[:10]
-
-
-def resolve_ingredient_ids(tokens):
-    if not tokens:
-        return {}
-    names = [t for t in tokens]
-    rows = query('SELECT id,name FROM ingredients WHERE name IN ('+','.join(['%s']*len(names))+')', names)
-    known = {r['name']: r['id'] for r in rows}
-    if len(known) == len(tokens):
-        return {t: known[t] for t in tokens}
-    all_names = [r['name'] for r in query('SELECT name FROM ingredients')]
-    resolved = {}
-    for t in tokens:
-        if t in known:
-            resolved[t] = known[t]
-            continue
-        cand = get_close_matches(t, all_names, n=1, cutoff=0.8)
-        if cand:
-            rid = query('SELECT id FROM ingredients WHERE name=%s', (cand[0],), one=True)['id']
-            resolved[t] = rid
-    return resolved
-
-def dish_name_hits(raw):
-    raw = raw.strip()
-    if not raw:
-        return []
-    rows = list(query(
-        'SELECT id,name,cuisine,time_min,difficulty,veg,spice_level,image_url '
-        'FROM dishes WHERE name LIKE %s '
-        'ORDER BY CASE WHEN name=%s THEN 0 ELSE 1 END, LENGTH(name) ASC LIMIT 5',
-        (f'%{raw}%', raw)
-    ))
-    for r in rows:
-        r['score'] = 999
-        r['hit_labels'] = ['name match']
-    return rows
-
-
-
-def _center_crop_ratio(img, rw=16, rh=9):
-    w, h = img.size
-    tr = rw / rh
-    cr = w / h
-    if cr > tr:
-        nw = int(h * tr)
-        x = (w - nw) // 2
-        return img.crop((x, 0, x + nw, h))
-    nh = int(w / tr)
-    y = (h - nh) // 2
-    return img.crop((0, y, w, y + nh))
-
-
-def save_image(fileobj, w=1280, h=720):
-    img = Image.open(fileobj.stream)
-    if img.mode not in ('RGB', 'L'):
-        img = img.convert('RGB')
-    img = _center_crop_ratio(img, 16, 9).resize((w, h), Image.LANCZOS)
-    name = f"{uuid.uuid4().hex}.webp"
-    path = os.path.join(UPLOAD_DIR, name)
-    img.save(path, format='WEBP', quality=85, method=6)
-    return f"/static/uploads/{name}"
 
 
 @app.context_processor
@@ -168,67 +23,6 @@ def url_helpers():
         return url_for('history', **args)
     return dict(url_for_history=url_for_history)
 
-def get_prefs(user_id=1):
-    row = query('SELECT * FROM preferences WHERE user_id=%s', (user_id,), one=True)
-    if not row:
-        execute('INSERT INTO preferences (user_id) VALUES (%s)', (user_id,))
-        row = query('SELECT * FROM preferences WHERE user_id=%s', (user_id,), one=True)
-    return row
-
-def get_cooldown_days(user_id=1):
-    prefs = get_prefs(user_id)
-    d = prefs.get('cooldown_days') or config.COOLDOWN_DAYS
-    return int(d)
-
-def pick_candidate(user_id=1):
-    today = date.today()
-    cd = get_cooldown_days(user_id)
-    pf_sql, pf_params = pref_filter_sql(user_id)
-    params = [user_id, user_id, today - timedelta(days=cd)] + pf_params
-    rows = query('SELECT d.*, ul.last_cooked_at FROM user_library ul JOIN dishes d ON d.id=ul.dish_id WHERE ul.user_id=%s AND ul.active=1 AND d.id NOT IN (SELECT dish_id FROM day_plan WHERE user_id=%s AND date >= %s)' + pf_sql, params)
-    if not rows:
-        params = [user_id] + pf_params
-        rows = query('SELECT d.*, ul.last_cooked_at FROM user_library ul JOIN dishes d ON d.id=ul.dish_id WHERE ul.user_id=%s AND ul.active=1' + pf_sql, params)
-    if not rows:
-        return None
-    def score(r):
-        lc = r.get('last_cooked_at')
-        base = 0 if lc is None else (today - lc).days
-        return base + random.random()
-    rows.sort(key=score, reverse=True)
-    return rows[0]
-
-
-def get_or_create_today_plan(user_id=1):
-    today_d = date.today()
-    existing = query('SELECT d.*, dp.is_override FROM day_plan dp JOIN dishes d ON d.id=dp.dish_id WHERE dp.user_id=%s AND dp.date=%s', (user_id, today_d), one=True)
-    if existing:
-        return existing
-    prefs = get_prefs(user_id)
-    if prefs.get('auto_suggestions', 1):
-        cand = pick_candidate(user_id)
-        if cand:
-            execute('INSERT INTO day_plan (user_id,date,dish_id,is_override) VALUES (%s,%s,%s,0) ON DUPLICATE KEY UPDATE dish_id=VALUES(dish_id), is_override=0', (user_id, today_d, cand['id']))
-            return query('SELECT d.*, dp.is_override FROM day_plan dp JOIN dishes d ON d.id=dp.dish_id WHERE dp.user_id=%s AND dp.date=%s', (user_id, today_d), one=True)
-    return pick_candidate(user_id)
-
-def alt_picks(exclude_id, user_id=1, limit=3):
-    today = date.today()
-    cd = get_cooldown_days(user_id)
-    pf_sql, pf_params = pref_filter_sql(user_id)
-    params = [user_id, exclude_id, user_id, today - timedelta(days=cd)] + pf_params
-    rows = query('SELECT d.*, ul.last_cooked_at FROM user_library ul JOIN dishes d ON d.id=ul.dish_id WHERE ul.user_id=%s AND ul.active=1 AND d.id<>%s AND d.id NOT IN (SELECT dish_id FROM day_plan WHERE user_id=%s AND date>=%s)' + pf_sql, params)
-    if not rows:
-        params = [user_id, exclude_id] + pf_params
-        rows = query('SELECT d.*, ul.last_cooked_at FROM user_library ul JOIN dishes d ON d.id=ul.dish_id WHERE ul.user_id=%s AND ul.active=1 AND d.id<>%s' + pf_sql, params)
-    random.shuffle(rows)
-    return rows[:limit]
-
-
-def days_ago(d):
-    if not d:
-        return None
-    return (date.today() - d).days
 
 @app.route('/')
 def today():
@@ -237,6 +31,7 @@ def today():
     recent = query('SELECT d.*, dp.date AS cooked_date FROM day_plan dp JOIN dishes d ON d.id=dp.dish_id WHERE dp.user_id=%s ORDER BY dp.date DESC LIMIT 6', (1,))
     cooldown = get_cooldown_days(1)
     return render_template('today.html', pick=pick, alts=alts, recent=recent, days_ago=days_ago, cooldown=cooldown, rotate_seconds=config.DEV_ROTATE_SECONDS)
+
 
 @app.get('/api/pick')
 def api_pick():
@@ -250,6 +45,7 @@ def api_pick():
         x['last_cooked_at'] = v.isoformat()
     return jsonify(x)
 
+
 @app.post('/cook')
 def cook():
     dish_id = int(request.form['dish_id'])
@@ -262,147 +58,18 @@ def cook():
     execute('UPDATE user_library SET last_cooked_at=%s WHERE user_id=%s AND dish_id=%s', (today_d, 1, dish_id))
     return redirect(url_for('today'))
 
+
 @app.post('/swap')
 def swap():
     dish_id = int(request.form['dish_id'])
     alts = alt_picks(dish_id, 1, 3)
     return jsonify(alts)
 
+
 @app.get('/override')
 def override_get():
     return render_template('override.html')
 
-
-def normalize_tokens(s):
-    s = s.lower().strip()
-    s = s.replace(',', ' ')
-    s = s.replace('/', ' | ')
-    s = re.sub(r'\band\b', '+', s)
-    s = re.sub(r'\bor\b', '|', s)
-    s = re.sub(r'\s+', ' ', s)
-    s = re.sub(r'([+\-|])', r' \1 ', s).strip()
-    parts = [p for p in s.split() if p]
-    op = '+'
-    required, optional, excluded = [], [], []
-    for p in parts:
-        if p in ['+','|','-']:
-            op = p
-            continue
-        if p.endswith('s') and len(p) > 3:
-            p = p[:-1]
-        aliases = {
-            'chkn':'chicken','chikn':'chicken','chk':'chicken','murghi':'chicken',
-            'qeema':'keema','keema':'keema',
-            'aloo':'potato','bhindi':'okra','saag':'spinach','machli':'fish',
-            'dal':'dal','daal':'dal','dāl':'dal',
-            'masoor':'masoor dal','moong':'moong dal','chanadal':'chana dal','chana':'chana dal','mash':'urad dal','urad':'urad dal',
-            'chawal':'rice','chāwal':'rice'
-        }
-
-        p = aliases.get(p, p)
-        if op == '-':
-            excluded.append(p)
-        elif op == '|':
-            optional.append(p)
-        else:
-            required.append(p)
-    return required, optional, excluded
-
-
-def resolve_ingredient_ids(tokens):
-    if not tokens:
-        return {}
-    names = [t for t in tokens]
-    rows = query('SELECT id,name FROM ingredients WHERE name IN ('+','.join(['%s']*len(names))+')', names)
-    known = {r['name']: r['id'] for r in rows}
-    if len(known) == len(tokens):
-        return {t: known[t] for t in tokens}
-    all_names = [r['name'] for r in query('SELECT name FROM ingredients')]
-    resolved = {}
-    for t in tokens:
-        if t in known:
-            resolved[t] = known[t]
-            continue
-        cand = get_close_matches(t, all_names, n=1, cutoff=0.8)
-        if cand:
-            rid = query('SELECT id FROM ingredients WHERE name=%s', (cand[0],), one=True)['id']
-            resolved[t] = rid
-    return resolved
-
-
-def dish_name_hits(raw):
-    raw = raw.strip()
-    if not raw:
-        return []
-    rows = query('SELECT id,name,cuisine,time_min,difficulty,veg,spice_level,image_url FROM dishes WHERE name LIKE %s ORDER BY CASE WHEN name=%s THEN 0 ELSE 1 END, LENGTH(name) ASC LIMIT 5', (f'%{raw}%', raw))
-    for r in rows:
-        r['score'] = 999
-        r['hit_labels'] = ['name match']
-    return rows
-
-
-def match_dishes(raw):
-    req, opt, ex = normalize_tokens(raw)
-    rid = resolve_ingredient_ids(req)
-    oid = resolve_ingredient_ids(opt)
-    xid = resolve_ingredient_ids(ex)
-    req_ids = set(rid.values())
-    opt_ids = set(oid.values())
-    ex_ids = set(xid.values())
-    all_ids = list(req_ids | opt_ids | ex_ids)
-
-    hits = list(dish_name_hits(raw))  # force list
-
-    if not all_ids:
-        return hits
-
-    placeholders = ','.join(['%s']*len(all_ids))
-    rows = query(f'''
-        SELECT d.id,d.name,d.cuisine,d.time_min,d.difficulty,d.veg,d.spice_level,d.image_url,
-               GROUP_CONCAT(i.id) AS hit_ids,
-               GROUP_CONCAT(i.name) AS hit_names
-        FROM dishes d
-        JOIN dish_ingredients di ON di.dish_id=d.id
-        JOIN ingredients i ON i.id=di.ingredient_id
-        WHERE di.ingredient_id IN ({placeholders})
-        GROUP BY d.id
-    ''', all_ids)
-
-    out, seen = [], set()
-    for r in rows:
-        ids = set(map(int, r['hit_ids'].split(','))) if r['hit_ids'] else set()
-        if ex_ids & ids:
-            continue
-        if req_ids and not req_ids.issubset(ids):
-            continue
-        score = len(req_ids & ids)*2 + len(opt_ids & ids)
-        if score <= 0:
-            continue
-        r['score'] = score
-        r['hit_labels'] = [n for n in (r['hit_names'] or '').split(',') if n]
-        if r['id'] not in seen:
-            out.append(r); seen.add(r['id'])
-
-    out.sort(key=lambda x: (-x['score'], x.get('time_min') or 999, x['name']))
-    id_hits = {h['id'] for h in hits}
-    return hits + [o for o in out if o['id'] not in id_hits]
-
-
-def ensure_ing(name):
-    r = query('SELECT id FROM ingredients WHERE name=%s', (name,), one=True)
-    return r['id'] if r else execute('INSERT INTO ingredients (name) VALUES (%s)', (name,))
-
-def ensure_dish(d):
-    r = query('SELECT id FROM dishes WHERE name=%s', (d['name'],), one=True)
-    if r:
-        return r['id']
-    return execute('INSERT INTO dishes (name,cuisine,time_min,difficulty,veg,spice_level,image_url) VALUES (%s,%s,%s,%s,%s,%s,%s)',
-                   (d['name'], d['cuisine'], d['time_min'], d['difficulty'], d['veg'], d['spice_level'], d.get('image_url','/static/img/placeholder.jpg')))
-
-def link_di(dish_id, ing_names):
-    for n in ing_names:
-        iid = ensure_ing(n)
-        execute('INSERT IGNORE INTO dish_ingredients (dish_id,ingredient_id) VALUES (%s,%s)', (dish_id, iid))
 
 @app.post('/admin/seed_pk_basics')
 def seed_pk_basics():
@@ -422,7 +89,6 @@ def seed_pk_basics():
     return redirect(url_for('library'))
 
 
-
 @app.post('/override')
 def override_post():
     raw = request.form.get('ingredients','').strip()
@@ -430,7 +96,6 @@ def override_post():
     web = web_find_recipes(raw)
     items = combine_override_results(raw, lib, web)
     return render_template('override_results.html', raw=raw, items=items, lib_count=len(lib), web_count=len(web))
-
 
 
 @app.post('/override/add')
@@ -452,6 +117,7 @@ def override_confirm():
     execute('UPDATE user_library SET last_cooked_at=%s WHERE user_id=%s AND dish_id=%s', (today_d, 1, dish_id))
     return redirect(url_for('today'))
 
+
 @app.get('/library')
 def library():
     q = request.args.get('q','').strip()
@@ -460,6 +126,7 @@ def library():
     else:
         rows = query('SELECT d.*, ul.last_cooked_at FROM user_library ul JOIN dishes d ON d.id=ul.dish_id WHERE ul.user_id=%s AND ul.active=1 ORDER BY d.name ASC', (1,))
     return render_template('library.html', rows=rows, q=q, days_ago=days_ago)
+
 
 @app.post('/library/add')
 def library_add():
@@ -555,6 +222,7 @@ def discover_add():
     execute('INSERT IGNORE INTO user_library (user_id,dish_id) VALUES (%s,%s)', (1,dish_id))
     return redirect(url_for('library'))
 
+
 @app.get('/settings')
 def settings():
     prefs = get_prefs(1)
@@ -566,6 +234,7 @@ def settings():
     vegp = query('SELECT ROUND(AVG(d.veg)*100) p FROM user_library ul JOIN dishes d ON d.id=ul.dish_id WHERE ul.user_id=%s AND ul.active=1', (1,), one=True)['p']
     stats['veg_pct'] = vegp or 0
     return render_template('settings.html', prefs=prefs, stats=stats)
+
 
 @app.post('/settings')
 def settings_save():
@@ -583,6 +252,7 @@ def settings_save():
     execute('INSERT INTO preferences (user_id,diet,spice_level,time_max,notify_time,daily_suggestions,weekly_discovery,auto_suggestions,cooldown_days,allergies,avoid,theme) VALUES (%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s) ON DUPLICATE KEY UPDATE diet=VALUES(diet),spice_level=VALUES(spice_level),time_max=VALUES(time_max),notify_time=VALUES(notify_time),daily_suggestions=VALUES(daily_suggestions),weekly_discovery=VALUES(weekly_discovery),auto_suggestions=VALUES(auto_suggestions),cooldown_days=VALUES(cooldown_days),allergies=VALUES(allergies),avoid=VALUES(avoid),theme=VALUES(theme)', (1,diet,spice,time_max,notify,daily,weekly,auto,cooldown_days,allergies,avoid,theme))
     return redirect(url_for('settings'))
 
+
 @app.get('/settings/export')
 def export_library():
     rows = query('SELECT d.name,d.cuisine,d.time_min,difficulty,veg,spice_level FROM user_library ul JOIN dishes d ON d.id=ul.dish_id WHERE ul.user_id=%s AND ul.active=1 ORDER BY d.name', (1,))
@@ -593,6 +263,7 @@ def export_library():
         w.writerow([r['name'], r['cuisine'] or '', r['time_min'] or '', r['difficulty'] or '', r['veg'] or 0, r['spice_level'] or ''])
     csv_data = output.getvalue()
     return Response(csv_data, mimetype='text/csv', headers={'Content-Disposition': 'attachment; filename=library.csv'})
+
 
 @app.post('/settings/import')
 def import_library():
@@ -632,69 +303,6 @@ def dish_image(dish_id):
         return redirect(url_for('library'))
     execute('UPDATE dishes SET image_url=%s WHERE id=%s', (url, dish_id))
     return redirect(url_for('library'))
-
-
-def pref_filter_sql(user_id=1):
-    p = get_prefs(user_id)
-    w, params = [], []
-
-    tm = p.get('time_max')
-    if tm:
-        w.append('d.time_min<=%s'); params.append(int(tm))
-
-    diet = (p.get('diet') or 'None')
-    if diet in ('Veg','Vegan'):
-        w.append('d.veg=1')
-
-    spice = (p.get('spice_level') or 'Medium')
-    if spice == 'Low':
-        w.append("COALESCE(d.spice_level,'Medium') NOT IN ('High','Spicy')")
-    elif spice == 'Medium':
-        w.append("COALESCE(d.spice_level,'Medium') NOT IN ('Spicy')")
-
-    avoid_tokens = []
-    for src in (p.get('allergies') or '', p.get('avoid') or ''):
-        r, o, x = normalize_tokens(src)
-        avoid_tokens.extend([t for t in (r + o + x) if t])
-
-    if avoid_tokens:
-        placeholders = ','.join(['%s'] * len(avoid_tokens))
-        w.append(
-            f"d.id NOT IN (SELECT di.dish_id FROM dish_ingredients di "
-            f"JOIN ingredients i ON i.id=di.ingredient_id "
-            f"WHERE i.name IN ({placeholders}))"
-        )
-        params.extend(avoid_tokens)
-
-    sql = (' AND ' + ' AND '.join(w)) if w else ''
-    return sql, params
-
-
-def ingredient_terms_from_text(s):
-    s = (s or '').lower().strip()
-    if not s: return []
-    parts = re.split(r'[,\+\|\-\/]+', s)
-    aliases = {
-        'chkn':'chicken','chikn':'chicken','chk':'chicken','murghi':'chicken',
-        'qeema':'keema','keema':'keema',
-        'aloo':'potato','bhindi':'okra','saag':'spinach','machli':'fish',
-        'dal':'dal','daal':'dal','dāl':'dal',
-        'masoor':'masoor dal','moong':'moong dal','chanadal':'chana dal','chana':'chana dal',
-        'mash':'urad dal','urad':'urad dal',
-        'chawal':'rice','chāwal':'rice'
-    }
-    out=[]
-    for p in parts:
-        t=p.strip()
-        if not t: continue
-        if t.endswith('s') and len(t)>3: t=t[:-1]
-        t=aliases.get(t,t)
-        out.append(t)
-    seen=set(); flat=[]
-    for t in out:
-        if t and t not in seen:
-            flat.append(t); seen.add(t)
-    return flat
 
 
 if __name__ == '__main__':
